# models/request-models/update-movie.go
package requestmodels

type UpdateMovie struct {
	Id          int    `json:"id"`
	Name        string `json:"name,omitempty"`
	Description string `json:"description,omitempty"`
	Type        string `json:"type,omitempty"`
}